Migrate Thread room action
