+++ apps/meteor/app/threads/client/index.ts
@@ -1,4 +1,3 @@
-import './flextab/threadlist';
 import './messageAction/follow';
 import './messageAction/unfollow';
 import './messageAction/replyInThread';
